Force installation of elasticsearch v5
The ElastAlert version that we're using by default comes with the v7 of the `elasticsearch` library, which fails to run aggregation rules because of incompatibility. This fixes the issue by pinning the elasticsearch library to v5 while we still use Elasticsearch 5. The `--force-reinstall` flag is necessary because of the hardcoded requirement in ElastAlert, it might trigger some warnings about incompatible version but can be safely ignored.